Special and physical attacks done

+++ src/src/main/java/Pokemon/Attack.java
@@ -71,7 +71,7 @@ public class Attack {
         double criticalProb = criticalProb(pokemon)/256;
         return rand < criticalProb;
     }
-    public float criticalDamage(Pokemon pokemon) {
-        return (float) (2 * pokemon.getLevel() + 5) / (pokemon.getLevel() + 5);
+    public double criticalDamage(Pokemon pokemon) {
+        return (double) (2 * pokemon.getLevel() + 5) / (pokemon.getLevel() + 5);
     }
 }
